Largely rewrite ./test/travis.sh
We now use the docker container to set up boulder, which subsumes most
of the logic in the script.

This also adds a (generous) timeout when waiting for boulder to start;
while attempting to get the old script to work, I hit a couple cases
where boulder was dying on startup, and Travis just never stopped.