Change setup.py metadata to point at my stuff
TODO: add an AUTHORS/CONTRIBUTORS file or somesuch. The metadata in
setup.py should point at the person actively maintaining it the code,
but I don't want to just take the original author's name off the thing
entirely, and there are other people's contributions in here as well.
It's all in the git history of course, but...