Drop support for Python 3.3
Python 3.3 has been EOL for five years now, and the CI has started
failing due to what seem to be changes in setuptools. Even CentOS/RHEL 6
has Python 3.4 available, so it doesn't seem worth struggling with.

Note that the system python on RHEL/CentOS 6 is still Python 2.6, so
that will probably continue to be supported for some time.